# admin/src/api/generated/models/DevLoginRequest.ts
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface DevLoginRequest
 */
export interface DevLoginRequest {
    /**
     * 
     * @type {string}
     * @memberof DevLoginRequest
     */
    openid?: string;
}

/**
 * Check if a given object implements the DevLoginRequest interface.
 */
export function instanceOfDevLoginRequest(value: object): value is DevLoginRequest {
    return true;
}

export function DevLoginRequestFromJSON(json: any): DevLoginRequest {
    return DevLoginRequestFromJSONTyped(json, false);
}

export function DevLoginRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): DevLoginRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'openid': json['openid'] == null ? undefined : json['openid'],
    };
}

export function DevLoginRequestToJSON(json: any): DevLoginRequest {
    return DevLoginRequestToJSONTyped(json, false);
}

export function DevLoginRequestToJSONTyped(value?: DevLoginRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'openid': value['openid'],
    };
}