# admin/src/api/generated/models/ReviewAppealRequest.ts
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ReviewAppealRequest
 */
export interface ReviewAppealRequest {
    /**
     * 
     * @type {string}
     * @memberof ReviewAppealRequest
     */
    appealType?: string;
    /**
     * 
     * @type {string}
     * @memberof ReviewAppealRequest
     */
    appealText?: string;
}

/**
 * Check if a given object implements the ReviewAppealRequest interface.
 */
export function instanceOfReviewAppealRequest(value: object): value is ReviewAppealRequest {
    return true;
}

export function ReviewAppealRequestFromJSON(json: any): ReviewAppealRequest {
    return ReviewAppealRequestFromJSONTyped(json, false);
}

export function ReviewAppealRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): ReviewAppealRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'appealType': json['appeal_type'] == null ? undefined : json['appeal_type'],
        'appealText': json['appeal_text'] == null ? undefined : json['appeal_text'],
    };
}

export function ReviewAppealRequestToJSON(json: any): ReviewAppealRequest {
    return ReviewAppealRequestToJSONTyped(json, false);
}

export function ReviewAppealRequestToJSONTyped(value?: ReviewAppealRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'appeal_type': value['appealType'],
        'appeal_text': value['appealText'],
    };
}