# admin/src/api/generated/models/SmsCodeResponse.ts
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface SmsCodeResponse
 */
export interface SmsCodeResponse {
    /**
     * 
     * @type {string}
     * @memberof SmsCodeResponse
     */
    message?: string;
    /**
     * 
     * @type {number}
     * @memberof SmsCodeResponse
     */
    expiresIn?: number;
}

/**
 * Check if a given object implements the SmsCodeResponse interface.
 */
export function instanceOfSmsCodeResponse(value: object): value is SmsCodeResponse {
    return true;
}

export function SmsCodeResponseFromJSON(json: any): SmsCodeResponse {
    return SmsCodeResponseFromJSONTyped(json, false);
}

export function SmsCodeResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): SmsCodeResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'message': json['message'] == null ? undefined : json['message'],
        'expiresIn': json['expires_in'] == null ? undefined : json['expires_in'],
    };
}

export function SmsCodeResponseToJSON(json: any): SmsCodeResponse {
    return SmsCodeResponseToJSONTyped(json, false);
}

export function SmsCodeResponseToJSONTyped(value?: SmsCodeResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'message': value['message'],
        'expires_in': value['expiresIn'],
    };
}