# openapi/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/v1/articles/{id}":{"get":{"tags":["article-controller"],"operationId":"getArticle","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ArticleDetailResponse"}}}}}},"put":{"tags":["article-controller"],"operationId":"updateDraft","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDraftRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreateDraftResponse"}}}}}}},"/api/v1/admin/categories/{id}":{"put":{"tags":["admin-category-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminCategoryRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AdminCategoryDto"}}}}}},"delete":{"tags":["admin-category-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/topics":{"get":{"tags":["topic-controller"],"operationId":"list","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string"}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"period","in":"query","required":false,"schema":{"type":"string"}},{"name":"mode","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TopicListResponse"}}}}}},"post":{"tags":["topic-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTopicRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreateTopicResponse"}}}}}}},"/api/v1/tags":{"post":{"tags":["tag-controller"],"operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTagRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TagItemDto"}}}}}}},"/api/v1/messages/notifications/read":{"post":{"tags":["notification-controller"],"operationId":"markRead","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MarkNotificationsReadRequest"}}}},"responses":{"200":{"description":"OK"}}}},"/api/v1/me/wechat-profile":{"post":{"tags":["me-controller"],"operationId":"updateWechatProfile","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WechatProfileRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserSummary"}}}}}}},"/api/v1/me/phone/bind":{"post":{"tags":["me-controller"],"operationId":"bindPhone","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BindPhoneRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/me/password":{"post":{"tags":["me-controller"],"operationId":"changePassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/me/account-name":{"post":{"tags":["me-controller"],"operationId":"changeAccountName","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangeAccountNameRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/auth/wechat/login":{"post":{"tags":["auth-controller"],"operationId":"wechatLogin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WechatLoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}}}},"/api/v1/auth/sms/code":{"post":{"tags":["auth-controller"],"operationId":"sendSmsCode","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SmsCodeRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SmsCodeResponse"}}}}}}},"/api/v1/auth/phone/login":{"post":{"tags":["auth-controller"],"operationId":"phoneLogin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhoneLoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}}}},"/api/v1/auth/password/login":{"post":{"tags":["auth-controller"],"operationId":"passwordLogin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordLoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}}}},"/api/v1/auth/dev/login":{"post":{"tags":["auth-controller"],"operationId":"devLogin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DevLoginRequest"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DevLoginResponse"}}}}}}},"/api/v1/auth/admin/login":{"post":{"tags":["auth-controller"],"operationId":"adminLogin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminLoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}}}},"/api/v1/articles/{id}/review-appeal":{"post":{"tags":["article-controller"],"operationId":"reviewAppeal","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewAppealRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReviewAppealResponse"}}}}}}},"/api/v1/articles/{id}/publish":{"post":{"tags":["article-controller"],"operationId":"publish","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PublishResponse"}}}}}}},"/api/v1/articles/drafts":{"post":{"tags":["article-controller"],"operationId":"createDraft","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDraftRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreateDraftResponse"}}}}}}},"/api/v1/admin/reviews/{id}/reject":{"post":{"tags":["admin-review-controller"],"operationId":"reject","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewRejectRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/admin/reviews/{id}/approve":{"post":{"tags":["admin-review-controller"],"operationId":"approve","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/admin/reports/{id}/resolve":{"post":{"tags":["admin-report-controller"],"operationId":"resolve","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportResolveRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/admin/features":{"get":{"tags":["admin-feature-controller"],"operationId":"list_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FeatureConfigDto"}}}}}}},"post":{"tags":["admin-feature-controller"],"operationId":"update_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeatureUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/admin/categories":{"get":{"tags":["admin-category-controller"],"operationId":"list_2","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AdminCategoryDto"}}}}}}},"post":{"tags":["admin-category-controller"],"operationId":"create_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminCategoryRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AdminCategoryDto"}}}}}}},"/api/v1/admin/announcements":{"get":{"tags":["admin-announcement-controller"],"operationId":"list_3","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AnnouncementDto"}}}}}}},"post":{"tags":["admin-announcement-controller"],"operationId":"create_3","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAnnouncementRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AnnouncementDto"}}}}}}},"/api/test/essay-list":{"post":{"tags":["test-essay-controller"],"operationId":"getList","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TestRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TestEssayDto"}}}}}}}},"/api/v1/topics/{id}":{"get":{"tags":["topic-controller"],"operationId":"detail","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TopicDetailResponse"}}}}}}},"/api/v1/topics/selectable":{"get":{"tags":["topic-controller"],"operationId":"selectable","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TopicListResponse"}}}}}}},"/api/v1/topics/ongoing":{"get":{"tags":["topic-controller"],"operationId":"ongoing","parameters":[{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":5}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TopicListResponse"}}}}}}},"/api/v1/tags/search":{"get":{"tags":["tag-controller"],"operationId":"search","parameters":[{"name":"q","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TagListResponse"}}}}}}},"/api/v1/rankings/hot":{"get":{"tags":["ranking-controller"],"operationId":"hot","parameters":[{"name":"window","in":"query","required":false,"schema":{"type":"string","default":"24h"}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HotRankingResponse"}}}}}}},"/api/v1/messages/notifications":{"get":{"tags":["notification-controller"],"operationId":"listNotifications","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":1}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NotificationListResponse"}}}}}}},"/api/v1/me":{"get":{"tags":["me-controller"],"operationId":"getMe","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MeResponse"}}}}}},"delete":{"tags":["me-controller"],"operationId":"deactivate","responses":{"200":{"description":"OK"}}}},"/api/v1/me/topics":{"get":{"tags":["me-controller"],"operationId":"myTopics","parameters":[{"name":"role","in":"query","required":false,"schema":{"type":"string","default":"created"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TopicListResponse"}}}}}}},"/api/v1/me/articles":{"get":{"tags":["me-controller"],"operationId":"listArticles","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":1}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MyArticlesResponse"}}}}}}},"/api/v1/me/account":{"get":{"tags":["me-controller"],"operationId":"accountSecurity","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccountSecurityResponse"}}}}}}},"/api/v1/config/client":{"get":{"tags":["config-controller"],"operationId":"client","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClientConfigResponse"}}}}}}},"/api/v1/categories/tree":{"get":{"tags":["category-controller"],"operationId":"tree","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CategoryTreeResponse"}}}}}}},"/api/v1/categories/preview":{"get":{"tags":["category-controller"],"operationId":"preview","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CategoryPreviewResponse"}}}}}}},"/api/v1/articles/featured":{"get":{"tags":["article-controller"],"operationId":"featured","parameters":[{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FeaturedArticlesResponse"}}}}}}},"/api/v1/admin/reviews":{"get":{"tags":["admin-review-controller"],"operationId":"list_4","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":1}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReviewListResponse"}}}}}}},"/api/v1/admin/reviews/{id}":{"get":{"tags":["admin-review-controller"],"operationId":"get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReviewDetailDto"}}}}}}},"/api/v1/admin/reports":{"get":{"tags":["admin-report-controller"],"operationId":"list_5","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReportDto"}}}}}}}}},"components":{"schemas":{"CreateDraftRequest":{"type":"object","properties":{"title":{"type":"string"},"content":{"type":"string"},"tag_ids":{"type":"array","items":{"type":"integer","format":"int64"}},"category_id":{"type":"integer","format":"int64"},"topic_id":{"type":"integer","format":"int64"},"cover_url":{"type":"string"},"image_urls":{"type":"array","items":{"type":"string"}}}},"CreateDraftResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"status":{"type":"string"},"word_count":{"type":"integer","format":"int32"},"created_at":{"type":"string","format":"date-time"}}},"AdminCategoryRequest":{"type":"object","properties":{"parent_id":{"type":"integer","format":"int64"},"name":{"type":"string"},"sort_order":{"type":"integer","format":"int32"},"is_home_recommended":{"type":"boolean"},"home_sort_order":{"type":"integer","format":"int32"}}},"AdminCategoryDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"parent_id":{"type":"integer","format":"int64"},"name":{"type":"string"},"is_home_recommended":{"type":"boolean"},"home_sort_order":{"type":"integer","format":"int32"}}},"CreateTopicRequest":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"duration_days":{"type":"integer","format":"int32"},"cover_url":{"type":"string"},"reward_pool_amount":{"type":"integer","format":"int32"},"reward_top_n":{"type":"integer","format":"int32"},"reward_ratio":{"type":"string"}}},"CreateTopicResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"status":{"type":"string"},"message":{"type":"string"}}},"CreateTagRequest":{"type":"object","properties":{"name":{"type":"string"}}},"TagItemDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"MarkNotificationsReadRequest":{"type":"object","properties":{"ids":{"type":"array","items":{"type":"integer","format":"int64"}}}},"WechatProfileRequest":{"type":"object","properties":{"nickname":{"type":"string"},"avatar_url":{"type":"string"}}},"UserSummary":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nickname":{"type":"string"},"avatar_url":{"type":"string"},"level":{"type":"string"},"level_name":{"type":"string"},"book_coin_balance":{"type":"integer","format":"int32"},"role":{"type":"string"}}},"BindPhoneRequest":{"type":"object","properties":{"phone":{"type":"string"},"code":{"type":"string"}}},"ChangePasswordRequest":{"type":"object","properties":{"account_name":{"type":"string"},"old_password":{"type":"string"},"new_password":{"type":"string"},"confirm_password":{"type":"string"}}},"ChangeAccountNameRequest":{"type":"object","properties":{"new_account_name":{"type":"string"},"password":{"type":"string"}}},"WechatLoginRequest":{"type":"object","properties":{"code":{"type":"string"}}},"LoginResponse":{"type":"object","properties":{"token":{"type":"string"},"expires_in":{"type":"integer","format":"int64"},"user":{"$ref":"#/components/schemas/UserSummary"}}},"SmsCodeRequest":{"type":"object","properties":{"phone":{"type":"string"},"purpose":{"type":"string"}}},"SmsCodeResponse":{"type":"object","properties":{"message":{"type":"string"},"expires_in":{"type":"integer","format":"int32"}}},"PhoneLoginRequest":{"type":"object","properties":{"phone":{"type":"string"},"code":{"type":"string"}}},"PasswordLoginRequest":{"type":"object","properties":{"account_name":{"type":"string"},"password":{"type":"string"}}},"DevLoginRequest":{"type":"object","properties":{"openid":{"type":"string"}}},"DevLoginResponse":{"type":"object","properties":{"token":{"type":"string"},"expires_in":{"type":"integer","format":"int64"},"user":{"$ref":"#/components/schemas/UserSummary"}}},"AdminLoginRequest":{"type":"object","properties":{"login_name":{"type":"string"},"password":{"type":"string"}}},"ReviewAppealRequest":{"type":"object","properties":{"appeal_type":{"type":"string"},"appeal_text":{"type":"string"}}},"ReviewAppealResponse":{"type":"object","properties":{"review_id":{"type":"integer","format":"int64"},"status":{"type":"string"},"message":{"type":"string"}}},"PublishResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"status":{"type":"string"},"published_at":{"type":"string","format":"date-time"},"reject_reason":{"type":"string"},"message":{"type":"string"}}},"ReviewRejectRequest":{"type":"object","properties":{"reason_code":{"type":"string"},"reason_text":{"type":"string"}}},"ReportResolveRequest":{"type":"object","properties":{"accepted":{"type":"boolean"}}},"FeatureConfigDto":{"type":"object","properties":{"key":{"type":"string"},"value":{"type":"string"}}},"FeatureUpdateRequest":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/FeatureConfigDto"}}}},"CreateAnnouncementRequest":{"type":"object","properties":{"title":{"type":"string"},"content":{"type":"string"},"publish_now":{"type":"boolean"}}},"AnnouncementDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"content":{"type":"string"},"published":{"type":"boolean"},"published_at":{"type":"string"},"created_at":{"type":"string"}}},"TestRequest":{"type":"object","properties":{"categoryId":{"type":"integer","format":"int32"},"authorName":{"type":"string"}}},"TestEssayDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"title":{"type":"string"}}},"TopicItemDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"description":{"type":"string"},"cover_url":{"type":"string"},"status":{"type":"string"},"article_count":{"type":"integer","format":"int32"},"view_count":{"type":"integer","format":"int32"},"start_at":{"type":"string"},"end_at":{"type":"string"},"days_remaining":{"type":"integer","format":"int32"}}},"TopicListResponse":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/TopicItemDto"}},"total":{"type":"integer","format":"int32"}}},"TopicDetailResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"description":{"type":"string"},"cover_url":{"type":"string"},"status":{"type":"string"},"article_count":{"type":"integer","format":"int32"},"view_count":{"type":"integer","format":"int32"},"duration_days":{"type":"integer","format":"int32"},"reward_pool_amount":{"type":"integer","format":"int32"},"start_at":{"type":"string"},"end_at":{"type":"string"},"creator_nickname":{"type":"string"},"days_remaining":{"type":"integer","format":"int32"}}},"TagListResponse":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/TagItemDto"}}}},"HotRankingItemDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"author_nickname":{"type":"string"},"view_count":{"type":"integer","format":"int32"},"like_count":{"type":"integer","format":"int32"},"hot_score":{"type":"number","format":"double"},"rank":{"type":"integer","format":"int32"}}},"HotRankingResponse":{"type":"object","properties":{"window":{"type":"string"},"items":{"type":"array","items":{"$ref":"#/components/schemas/HotRankingItemDto"}}}},"NotificationItemDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"type":{"type":"string"},"title":{"type":"string"},"content":{"type":"string"},"ref_type":{"type":"string"},"ref_id":{"type":"integer","format":"int64"},"is_read":{"type":"boolean"},"created_at":{"type":"string","format":"date-time"}}},"NotificationListResponse":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/NotificationItemDto"}},"total":{"type":"integer","format":"int64"},"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"}}},"MeResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nickname":{"type":"string"},"avatar_url":{"type":"string"},"level":{"type":"string"},"level_name":{"type":"string"},"book_coin_balance":{"type":"integer","format":"int32"},"follower_count":{"type":"integer","format":"int32"},"following_count":{"type":"integer","format":"int32"},"article_count":{"type":"integer","format":"int32"},"role":{"type":"string"},"word_limit":{"type":"integer","format":"int32"}}},"MyArticleItem":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"status":{"type":"string"},"updated_at":{"type":"string","format":"date-time"}}},"MyArticlesResponse":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/MyArticleItem"}},"total":{"type":"integer","format":"int64"},"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"}}},"AccountSecurityResponse":{"type":"object","properties":{"phone_masked":{"type":"string"},"phone_verified":{"type":"boolean"},"password_set":{"type":"boolean"},"wechat_bound":{"type":"boolean"},"account_name":{"type":"string"},"account_name_set":{"type":"boolean"},"can_change_account_name":{"type":"boolean"},"account_name_next_change_at":{"type":"string"}}},"ClientConfigResponse":{"type":"object","properties":{"recharge_enabled":{"type":"boolean"},"recharge_custom_amount_enabled":{"type":"boolean"}}},"CategoryNodeDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"children":{"type":"array","items":{"$ref":"#/components/schemas/CategoryNodeDto"}}}},"CategoryTreeResponse":{"type":"object","properties":{"categories":{"type":"array","items":{"$ref":"#/components/schemas/CategoryNodeDto"}}}},"CategoryPreviewArticleDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"cover_url":{"type":"string"},"hot_score":{"type":"number","format":"double"}}},"CategoryPreviewItemDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"articles":{"type":"array","items":{"$ref":"#/components/schemas/CategoryPreviewArticleDto"}}}},"CategoryPreviewResponse":{"type":"object","properties":{"categories":{"type":"array","items":{"$ref":"#/components/schemas/CategoryPreviewItemDto"}}}},"ArticleDetailResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"content":{"type":"string"},"status":{"type":"string"},"reject_reason":{"type":"string"},"ai_quality_score":{"type":"integer","format":"int32"},"tags":{"type":"array","items":{"$ref":"#/components/schemas/TagDto"}},"latest_review":{"$ref":"#/components/schemas/LatestReviewDto"},"published_at":{"type":"string","format":"date-time"},"created_at":{"type":"string","format":"date-time"}}},"LatestReviewDto":{"type":"object","properties":{"review_type":{"type":"string"},"ai_score":{"type":"integer","format":"int32"},"completed_at":{"type":"string","format":"date-time"}}},"TagDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"FeaturedArticleItemDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"cover_url":{"type":"string"},"author_nickname":{"type":"string"},"category_name":{"type":"string"},"featured_score":{"type":"number","format":"double"},"published_at":{"type":"string"}}},"FeaturedArticlesResponse":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/FeaturedArticleItemDto"}}}},"ReviewListItemDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"article_id":{"type":"integer","format":"int64"},"title":{"type":"string"},"review_type":{"type":"integer","format":"int32"},"priority":{"type":"integer","format":"int32"},"created_at":{"type":"string","format":"date-time"}}},"ReviewListResponse":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/ReviewListItemDto"}},"total":{"type":"integer","format":"int64"},"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"}}},"ReviewDetailDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"article_id":{"type":"integer","format":"int64"},"title":{"type":"string"},"content":{"type":"string"},"review_type":{"type":"integer","format":"int32"},"appeal_text":{"type":"string"},"ai_score":{"type":"integer","format":"int32"},"ai_dimensions":{"type":"string"},"created_at":{"type":"string","format":"date-time"}}},"ReportDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"reporter_id":{"type":"integer","format":"int64"},"target_type":{"type":"integer","format":"int32"},"target_id":{"type":"integer","format":"int64"},"reason":{"type":"string"},"reason_detail":{"type":"string"},"status":{"type":"integer","format":"int32"},"created_at":{"type":"string"}}}}}}